refactor: extract _get_labels_names into utils
