+++ tests/test_analysis/test_decomposition/test_plots.py
@@ -3,7 +3,6 @@ from sklearn.linear_model import LinearRegression
 from sklearn.linear_model import TheilSenRegressor
 
 from etna.analysis import plot_trend
-from etna.analysis.decomposition.plots import _get_labels_names
 from etna.transforms import ChangePointsTrendTransform
 from etna.transforms import LinearTrendTransform
 from etna.transforms import STLTransform
@@ -31,23 +30,3 @@ def test_plot_bin_seg(example_tsdf, detrend_model):
 @pytest.mark.parametrize("period", (7, 30))
 def test_plot_stl(example_tsdf, period):
     plot_trend(ts=example_tsdf, trend_transform=STLTransform(in_column="target", period=period))
-
-
-@pytest.mark.parametrize(
-    "poly_degree, expect_values, trend_class",
-    (
-        [1, True, LinearTrendTransform],
-        [2, False, LinearTrendTransform],
-        [1, True, TheilSenTrendTransform],
-        [2, False, TheilSenTrendTransform],
-    ),
-)
-def test_get_labels_names_linear_coeffs(example_tsdf, poly_degree, expect_values, trend_class):
-    ln_tr = trend_class(in_column="target", poly_degree=poly_degree)
-    ln_tr.fit_transform(example_tsdf)
-    segments = example_tsdf.segments
-    _, linear_coeffs = _get_labels_names([ln_tr], segments)
-    if expect_values:
-        assert list(linear_coeffs.values()) != ["", ""]
-    else:
-        assert list(linear_coeffs.values()) == ["", ""]
